+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - "You miss 100 percent of the chances you don't take. — Wayne Gretzky" — Michael Scott
 
-- [apm] fix: Check for performance.timing in webworkers #2491
+- [apm] fix: Check for performance.timing in webworkers (#2491)
 
 ## 5.14.0