+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [apm] fix: Use Performance API for timings when available, including Web Workers (#2492)
 - [apm] fix: Remove Performance references (#2495)
+- [apm] fix: Set `op` in node http.server transaction (#2496)
 
 ## 5.14.1