+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - "You miss 100 percent of the chances you don't take. — Wayne Gretzky" — Michael Scott
 
+## 5.14.2
+
 - [apm] fix: Use Performance API for timings when available, including Web Workers (#2492)
 - [apm] fix: Remove Performance references (#2495)
 - [apm] fix: Set `op` in node http.server transaction (#2496)